successfully integrated detail story with api

+++ app/src/main/java/com/zaghy/storyapp/detailstory/model/MResponseDetailStory.kt
@@ -29,11 +29,11 @@ data class Story(
 	val description: String? = null,
 
 	@field:SerializedName("lon")
-	val lon: Int? = null,
+	val lon: Double? = null,
 
 	@field:SerializedName("id")
 	val id: String? = null,
 
 	@field:SerializedName("lat")
-	val lat: Int? = null
+	val lat: Double? = null
 )
